refactor(ai-stream): remove steer-queue split-drift (review #1,#2,#5-doc,#8,#9,#14)
Dead code, stale docs/comments, and tests describing feat/chat-page's end-state
that this slice doesn't ship:

- Blocker 4: rewrite stream-manager.md (+ sibling ai docs) — the Steering
  section, send() contract, flow table, and edge cases described the removed
  abort-and-restart mechanism; now document enqueue/yield/steer-continuation.
- Blocker 5: delete the orphaned abortAndAwait method + its tests; the
  PersistentChatContextProvider describe block is renamed off the dead flow.
- #1: delete the turnId pipeline (ActiveStream.turnId, nextStreamTurnSequence,
  the ChatStreamLifecycle write, TopicStatusSnapshotEntry.turnId) — zero readers.
- #2: delete the dead AiChatRequestBody.knowledgeBaseIds contract field.
- #8: restore the per-row traceId assertion (this branch persists per-model
  traceIds, not a container trace) + an accurate startTurnRootSpans doc.
- #9: trim comments that described truth-branch-only code (steerReminder,
  steerYield, reservedMessages, the send() empty-models branch).
- #14: startingNextChat -> startingNextChatTopicIds; rename
  reasoningTimingTransform.ts -> withReasoningTimingMetadata.ts to match export.

typecheck:node + typecheck:web clean; 182 streamManager/idle tests pass.

(cherry picked from commit e9cc8676bd6615f6c42f81348b0d45d4e7161db1)
